Move code from SqlTransaction.Common into the class
