+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/SqlTransaction.Common.cs
@@ -12,130 +12,18 @@ namespace Microsoft.Data.SqlClient
     /// <include file='../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlTransaction.xml' path='docs/members[@name="SqlTransaction"]/SqlTransaction/*' />
     public sealed partial class SqlTransaction : DbTransaction
     {
-        private static int s_objectTypeCount; // EventSource Counter
-        internal readonly int _objectID = System.Threading.Interlocked.Increment(ref s_objectTypeCount);
-        internal readonly IsolationLevel _isolationLevel = IsolationLevel.ReadCommitted;
 
-        private SqlInternalTransaction _internalTransaction;
-        private readonly SqlConnection _connection;
-
-        private bool _isFromAPI;
-
-        internal SqlTransaction(SqlInternalConnection internalConnection, SqlConnection con,
-                                IsolationLevel iso, SqlInternalTransaction internalTransaction)
-        {
-#if NETFRAMEWORK
-            SqlConnection.VerifyExecutePermission();
-#endif
-            _isolationLevel = iso;
-            _connection = con;
-
-            if (internalTransaction == null)
-            {
-                _internalTransaction = new SqlInternalTransaction(internalConnection, TransactionType.LocalFromAPI, this);
-            }
-            else
-            {
-                Debug.Assert(internalConnection.CurrentTransaction == internalTransaction, "Unexpected Parser.CurrentTransaction state!");
-                _internalTransaction = internalTransaction;
-                _internalTransaction.InitParent(this);
-            }
-        }
-
-        ////////////////////////////////////////////////////////////////////////////////////////
-        // PROPERTIES
-        ////////////////////////////////////////////////////////////////////////////////////////
-
-        /// <include file='../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlTransaction.xml' path='docs/members[@name="SqlTransaction"]/Connection/*' />
-        public new SqlConnection Connection
-        {// MDAC 66655
-            get
-            {
-                if (IsZombied)
-                {
-                    return null;
-                }
-                else
-                {
-                    return _connection;
-                }
-            }
-        }
-
-        /// <include file='../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlTransaction.xml' path='docs/members[@name="SqlTransaction"]/DbConnection/*' />
-        protected override DbConnection DbConnection => Connection;
-
-        internal SqlInternalTransaction InternalTransaction => _internalTransaction;
-
-        /// <include file='../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlTransaction.xml' path='docs/members[@name="SqlTransaction"]/IsolationLevel/*' />
-        public override IsolationLevel IsolationLevel
-        {
-            get
-            {
-                ZombieCheck();
-                return _isolationLevel;
-            }
-        }
-
-        private bool Is2005PartialZombie => _internalTransaction !=null && _internalTransaction.IsCompleted;
-
-        internal bool IsZombied => _internalTransaction == null || _internalTransaction.IsCompleted;
-
-        internal int ObjectID => _objectID;
-
-        internal SqlStatistics Statistics
-        {
-            get
-            {
-                if (_connection != null)
-                {
-                    if (_connection.StatisticsEnabled)
-                    {
-                        return _connection.Statistics;
-                    }
-                }
-                return null;
-            }
-        }
 
         ////////////////////////////////////////////////////////////////////////////////////////
         // INTERNAL METHODS
         ////////////////////////////////////////////////////////////////////////////////////////
 
-        internal void Zombie()
-        {
-            // For Yukon, we have to defer "zombification" until
-            //                 we get past the users' next rollback, else we'll
-            //                 throw an exception there that is a breaking change.
-            //                 Of course, if the connection is already closed,
-            //                 then we're free to zombify...
-            SqlInternalConnection internalConnection = (_connection.InnerConnection as SqlInternalConnection);
-            if (internalConnection != null && !_isFromAPI)
-            {
-                SqlClientEventSource.Log.TryAdvancedTraceEvent("SqlTransaction.Zombie | ADV | Object Id {0} yukon deferred zombie", ObjectID);
-            }
-            else
-            {
-                _internalTransaction = null; // pre SQL 2005 zombification
-            }
-        }
+
 
         ////////////////////////////////////////////////////////////////////////////////////////
         // PRIVATE METHODS
         ////////////////////////////////////////////////////////////////////////////////////////
 
-        private void ZombieCheck()
-        {
-            // If this transaction has been completed, throw exception since it is unusable.
-            if (IsZombied)
-            {
-                if (Is2005PartialZombie)
-                {
-                    _internalTransaction = null; // SQL 2005 zombification
-                }
 
-                throw ADP.TransactionZombied(this);
-            }
-        }
     }
 }
